+++ src/settings/StaffSlips/StaffSlipForm.js
@@ -48,8 +48,8 @@ const StaffSlipForm = (props) => {
     : formatMessage({ id: 'ui-circulation.settings.staffSlips.new' });
 
   const footerPaneProps = {
-    pristine,
-    submitting,
+    isSaveButtonDisabled: pristine || submitting,
+    isSaveButtonAvailable: stripes.hasPerm('ui-circulation.settings.edit-staff-slips'),
     onCancel,
   };